[Go] add support GCP client options
Allow users to pass client options to the GCP clients for the
googleai and vertexai plugins.

To allow this and also to enable additional config in the future
without introducing breaking changes, add a Config struct to
each plugin, and have Init take it as an argument.

Also, slightly rewrite the samples to avoid redundant code for
looking up environment variables.